requirements.txt
# Web scraping
playwright==1.44.0

# HTTP & Telegram notifications
httpx==0.27.0

# OpenAI LLM scoring
openai==1.30.5

# Environment variable loading
python-dotenv==1.0.1

# Logging (built-in, no install needed)
# sqlite3 (built-in, no install needed)
